backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
pydantic[email]==2.9.2
python-dotenv==1.0.1
gunicorn==22.0.0
